+++ navigator/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version `0.1-alpha32` _(2021-10-20)_
+
+- **Breaking Change**: `getParentNavigator()` & `getSimpleNavigator()` have become property types.
+
 ## Version `0.1-alpha30` _(2021-08-03)_
 
 - **Fixed**: `IllegalStateException` when circularTransform runs before container's root view is laid out or is in layout.